adding/amending token auth for new corgi
